feat: 유저 JWT 토큰 구현 및 Users 기능 일부 구현 완료

+++ src/main/groovy/com/newsfeed/domain/users/service/UserService.java
@@ -1,5 +1,6 @@
 package com.newsfeed.domain.users.service;
 
+import com.newsfeed.common.Const;
 import com.newsfeed.common.dto.response.MessageResponse;
 import com.newsfeed.common.exception.ApplicationException;
 import com.newsfeed.config.PasswordEncoder;
@@ -7,6 +8,9 @@ import com.newsfeed.domain.followers.repository.FollowerRepository;
 import com.newsfeed.domain.users.dto.response.UserProfileResponseDto;
 import com.newsfeed.domain.users.entity.User;
 import com.newsfeed.domain.users.repository.UserRepository;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -34,31 +38,25 @@ public class UserService {
         return new UserProfileResponseDto(user, 0L, 0L);
     }
 
+    // 유저 로그인
+    public UserProfileResponseDto login(String email, String password) {
+        User findUser = findByEmailOrThrow(email);
+
+        if (!passwordEncoder.matches(password, findUser.getPassword())) {
+            throw new ApplicationException("비밀번호가 일치하지 않습니다", HttpStatus.UNAUTHORIZED);
+        }
+
+        return new UserProfileResponseDto(findUser, 0L, 0L);
+    }
+
     // 유저 단건 조회
     public UserProfileResponseDto findById(Long id) {
-        //JWT에서 현재 로그인한 유저 정보 가져오기
-
-
         // 조회하려는 유저 찾기
-        Optional<User> optionalUser = userRepository.findById(id);
-        if (optionalUser.isEmpty()) {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Does not exist id = " + id);
-        }
-        User findUser = optionalUser.get();
-
-        // 현재 로그인한 유저와 조회하려는 유저 같은지 비교
-
-        // 다르면 널처리
-
-        // 팔로워 & 팔로우 수 조회
-        Long followerCount = followerRepository.countByFollower(findUser);
-        Long followingCount = followerRepository.countByFollowing(findUser);
-
-        return new UserProfileResponseDto(findUser, followerCount, followingCount);
-
+        User findUser = findByIdOrElseThrow(id);
+        return new UserProfileResponseDto(findUser, 0L, 0L);
     }
-    // 유저 비밀번호 수정
 
+    // 유저 비밀번호 수정
     public void updatePassword(String oldPassword, String newPassword) {
         //JWT에서 내 로그인 정보 가져오기
 
@@ -78,24 +76,34 @@ public class UserService {
     }
 
     @Transactional
-    public MessageResponse deleteUser(Long userId, String password) {
+    public void deleteUser(Long userId, String password) {
         User findUser = findByIdOrElseThrow(userId);
+
+        if (findUser.getDeletedAt() != null) {
+            throw new ApplicationException("이미 탈퇴한 사용자입니다.", HttpStatus.BAD_REQUEST);
+        }
 
         if (!passwordEncoder.matches(password, findUser.getPassword())) {
             throw new ApplicationException("비밀번호가 일치하지 않습니다.", HttpStatus.UNAUTHORIZED);
         }
 
         findUser.deleteUser();
-        return MessageResponse.toDto("회원탈퇴에 성공했습니다.");
     }
 
+    // 해당 이메일을 가진 유저가 있는지 조회
     public boolean existsByEmail(String email) {
         return userRepository.existsByEmail(email);
     }
 
+    // Service 레벨에서 NULL 체크(유저 ID)
     public User findByIdOrElseThrow(Long userId) {
         return userRepository.findById(userId)
                 .orElseThrow(() -> new ApplicationException("해당 아이디와 일치하는 유저가 없습니다. id = " + userId, HttpStatus.NOT_FOUND));
+    }
 
+    // Service 레벨에서 NULL 체크(유저 이메일)
+    public User findByEmailOrThrow(String email) {
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new ApplicationException("해당 이메일과 일치하는 유저가 없습니다. email = " + email, HttpStatus.UNAUTHORIZED));
     }
 }
